requirements.txt
annotated-types==0.6.0
bidict==0.22.1
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
openai==1.10.0
pydantic==2.6.0
pydantic_core==2.16.1
python-dotenv==1.0.1
python-engineio==4.8.2
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
twilio==8.12.0
retell-sdk==3.5.0
fastapi==0.100.1
uvicorn==0.21.1
python-multipart==0.0.9
pandas